Separate `intro` and `photo`
- into 2 directories
- Move `upload` into `photo`
