+++ routes.js
@@ -7,19 +7,19 @@ app.use('/prototype/overseas', require('./routes/prototype/overseas'));
 app.use('/prototype/overseas-not-eligible', require('./routes/prototype/overseas-not-eligible'));
 app.use('/prototype/overseas-first', require('./routes/prototype/overseas-first'));
 app.use('/prototype/overseas-lost-change', require('./routes/prototype/overseas-lost-change'));
-app.use('/prototype/uploadphoto', require('./routes/prototype/uploadphoto'));
-app.use('/prototype/uploadphoto-oix', require('./routes/prototype/uploadphoto-oix'));
-app.use('/prototype/uploadphoto-child', require('./routes/prototype/uploadphoto-child'));
-app.use('/prototype/apply', require('./routes/prototype/apply'));
-app.use('/prototype/intro', require('./routes/prototype/intro'));
-app.use('/prototype/filter', require('./routes/prototype/filter'));
 app.use('/prototype/startpage', require('./routes/prototype/startpage'));
 app.use('/prototype/startpage-overseas', require('./routes/prototype/startpage-overseas'));
+app.use('/prototype/filter', require('./routes/prototype/filter'));
+app.use('/prototype/apply', require('./routes/prototype/apply'));
+app.use('/prototype/intro', require('./routes/prototype/intro'));
+app.use('/prototype/photo', require('./routes/prototype/photo'));
 app.use('/prototype/help', require('./routes/prototype/help'));
 app.use('/prototype/photo-guide', require('./routes/prototype/photo-guide'));
 app.use('/prototype/photo-booth-shop', require('./routes/prototype/photo-booth-shop'));
+app.use('/prototype/uploadphoto', require('./routes/prototype/uploadphoto'));
+app.use('/prototype/uploadphoto-oix', require('./routes/prototype/uploadphoto-oix'));
+app.use('/prototype/uploadphoto-child', require('./routes/prototype/uploadphoto-child'));
 app.use('/prototype/takephoto', require('./routes/prototype/takephoto'));
-app.use('/prototype/upload', require('./routes/prototype/upload'));
 app.use('/prototype/rejectedphoto', require('./routes/prototype/rejectedphoto'));
 app.use('/prototype/filter-common-temp', require('./routes/prototype/filter-common-temp'));
 app.use('/prototype/startpage-temp', require('./routes/prototype/startpage-temp'));
