+++ routes/prototype/intro/fields.js
@@ -4,69 +4,5 @@ module.exports = {
         validate: [
             'required'
         ]
-    },
-    'choose-photo': {
-        legend: {
-            value: 'What are you trying to do?',
-            className: 'visuallyhidden'
-        },
-        options: [{
-            value: 'myself',
-            label: 'Show me how to do it myself'
-        }, {
-            value: 'shop',
-            label: 'Show me how to get it from a shop (not yet prototyped)'
-        }, {
-            value: 'code',
-            label: 'I have a printed photo with a code'
-        }, {
-            value: 'upload',
-            label: 'I already have a digital photo'
-        }],
-        validate: [
-            'required',
-            {
-                type: 'equal',
-                arguments: ['upload', 'shop', 'code'],
-                redirect: '/../photo-guide'
-            }, {
-                type: 'equal',
-                arguments: ['upload', 'myself', 'code'],
-                redirect: '/../photo-booth-shop'
-            }, {
-                type: 'equal',
-                arguments: ['upload', 'upload', 'code'],
-                redirect: '/../upload'
-            }, {
-                type: 'equal',
-                arguments: ['upload', 'myself', 'shop'],
-                redirect: '/get-photo-code'
-            }
-        ]
-    },
-    'choose-photo-overseas': {
-        legend: {
-            value: 'What are you trying to do?',
-            className: 'visuallyhidden'
-        },
-        options: [{
-            value: 'myself',
-            label: 'Show me how to do it myself'
-        }, {
-            value: 'upload',
-            label: 'I already have a digital photo'
-        }],
-        validate: [
-            'required',
-            {
-                type: 'equal',
-                arguments: ['upload', 'shop'],
-                redirect: '/../photo-guide'
-            }, {
-                type: 'equal',
-                arguments: ['upload', 'upload'],
-                redirect: '/../upload'
-            }
-        ]
     }
 };
