--- routes/prototype/upload/index.js
@@ -1,16 +1,0 @@
-var app = require('express')(),
-    wizard = require('hmpo-form-wizard'),
-    steps = require('./steps'),
-    fields = require('./fields');
-
-app.use(require('hmpo-template-mixins')(fields, {
-    sharedTranslationKey: 'prototype'
-}));
-
-app.use(wizard(steps, fields, {
-    templatePath: 'prototype/upload',
-    name: 'common',
-    params: '/:action?'
-}));
-
-module.exports = app;
