+++ views/prototype/photo/choose-photo-method-old.html
@@ -1,0 +1,74 @@
+{{< layout}}
+
+{{$pageTitle}}Get your digital photo{{/pageTitle}}
+
+{{$header}}
+    <h1>Get a digital passport photo</h1>
+{{/header}}
+
+<!-- content changes made to you-need-a-photo.html -->
+{{$main-content}}
+<div class="grid-row" style="margin:15px;">
+<h1>Get a digital photo</h1>
+
+    {{< hmpo-partials-form}}
+    {{$form}}
+    <div class="grid-row">
+
+            <div class="column-one-third">
+              <!-- nothing here but keep it -->
+            </div>
+            <div class="column-full">
+                <a href="../photo-guide" class="guide-box">
+                        <img src="../../public/images/photo-guide-myself.png" alt="Your uploaded photo" title="Your uploaded photo" style="width:25.2%">
+                        <h3 style="margin-top:0;">How to take a digital photo</h3>
+                        <p>A step-by-step guide to taking a good passport photo. You'll need a friend to help you.</p>
+                </a>
+                <a href="../photo-booth-shop" class="guide-box last">
+                        <img src="../../public/images/photo-guide-shop.png" alt="Your uploaded photo" title="Your uploaded photo" width="200">
+                        <h3 style="margin-top:0;">If you go to a shop or booth</h3>
+                        <p>Check what to ask for at a photo shop and what you'll get from a booth.</p>
+                </a>
+      </div>
+</div>
+      <br/>
+
+      <div class="grid-row">
+          <div class="column-two-thirds" >
+            <h2>If you have a digital photo</h2>
+            <div id="choose-photo-group" class="form-group{{#errors.choose-photo}} error{{/errors.choose-photo}}">
+                <fieldset role="radiogroup" aria-required="upload">
+                  <!--
+                    <legend>
+                        <span class="form-label"></span>
+                        {{#errors.choose-photo}}<span id="{{key}}-error" class="error-message">{{errors.choose-photo.message}}</span>{{/errors.choose-photo}}
+                    </legend>
+                    <label class="block-label" for="choose-photo-myself">
+                        <input type="radio" name="choose-photo" id="choose-photo-myself" value="myself" aria-required="upload"{{#selected}}choose-photo=myself{{/selected}}{{#errors.choose-photo}} aria-describedby="{{key}}-error" aria-invalid="upload"{{/errors.choose-photo}}>
+                        I’ll take a digital photo &ndash; show me how
+                    </label>
+                    <label class="block-label" for="choose-photo-shop">
+                        <input type="radio" name="choose-photo" id="choose-photo-shop" value="shop" aria-required="upload"{{#selected}}choose-photo=shop{{/selected}}{{#errors.choose-photo}} aria-describedby="{{key}}-error" aria-invalid="upload"{{/errors.choose-photo}}>
+                        I’ll get one from a shop &ndash; tell me what to ask for
+                    </label>-->
+                   <label class="block-label" for="choose-photo-code">
+                        <input type="radio" name="choose-photo" id="choose-photo-code" value="code" aria-required="upload"{{#selected}}choose-photo=code{{/selected}}{{#errors.choose-photo}} aria-describedby="{{key}}-error" aria-invalid="upload"{{/errors.choose-photo}}>
+                        I've been to a shop or a booth and I have a code
+                    </label>
+                    <div style="clear:both"></div>
+                    <label class="block-label" for="choose-photo-upload">
+                        <input type="radio" name="choose-photo" id="choose-photo-upload" value="upload" aria-required="upload"{{#selected}}choose-photo=upload{{/selected}}{{#errors.choose-photo}} aria-describedby="{{key}}-error" aria-invalid="upload"{{/errors.choose-photo}}>
+                        I have a digital photo to upload
+                    </label>
+                </fieldset>
+            </div>
+          <!--  <p style="margin-top:-30px"><a href="../upload" >I have a digital photo</a></p>-->
+            {{#input-submit}}{{/input-submit}}
+      </div>
+      </div>
+        {{/form}}
+    {{/ hmpo-partials-form}}
+</div>
+{{/main-content}}
+
+{{/ layout}}
