+++ views/prototype/photo/choose-photo-method-overseas.html
@@ -1,0 +1,45 @@
+{{< layout}}
+
+{{$pageTitle}}Get your digital photo{{/pageTitle}}
+
+{{$header}}
+    <h1>Get your digital passport photo</h1>
+{{/header}}
+
+<!-- content changes made to you-need-a-photo.html -->
+{{$content}}
+
+    {{< hmpo-partials-form}}
+    {{$form}}
+            <p>Your photo must be taken in the last month and meet the <a href="/../priority_service_170301/photoguide-static/photorules" target="_blank">rules for digital passport photos</a>.</p>
+            <div id="choose-photo-group" class="form-group{{#errors.choose-photo}} error{{/errors.choose-photo}}">
+                <fieldset role="radiogroup" aria-required="upload">
+                    <legend>
+                        <span class="form-label"></span>
+                        {{#errors.choose-photo}}<span id="{{key}}-error" class="error-message">{{errors.choose-photo.message}}</span>{{/errors.choose-photo}}
+                    </legend>
+                    <label class="block-label" for="choose-photo-myself">
+                        <input type="radio" name="choose-photo" id="choose-photo-myself" value="myself" aria-required="upload"{{#selected}}choose-photo=myself{{/selected}}{{#errors.choose-photo}} aria-describedby="{{key}}-error" aria-invalid="upload"{{/errors.choose-photo}}>
+                        I’ll take a digital photo &ndash; show me how
+                    </label>
+                    <label class="block-label" for="choose-photo-shop">
+                        <input type="radio" name="choose-photo" id="choose-photo-shop" value="shop" aria-required="upload"{{#selected}}choose-photo=shop{{/selected}}{{#errors.choose-photo}} aria-describedby="{{key}}-error" aria-invalid="upload"{{/errors.choose-photo}}>
+                        I’ll get one from a shop &ndash; tell me what to ask for
+                    </label>
+                    <div style="clear:both"></div>
+                    <label class="block-label" for="choose-photo-upload">
+                        <input type="radio" name="choose-photo" id="choose-photo-upload" value="upload" aria-required="upload"{{#selected}}choose-photo=upload{{/selected}}{{#errors.choose-photo}} aria-describedby="{{key}}-error" aria-invalid="upload"{{/errors.choose-photo}}>
+                        I have a digital photo &ndash; upload it now
+                    </label>
+
+                </fieldset>
+            </div>
+          <!--  <p style="margin-top:-30px"><a href="../upload" >I have a digital photo</a></p>-->
+            {{#input-submit}}{{/input-submit}}
+
+        {{/form}}
+    {{/ hmpo-partials-form}}
+
+{{/content}}
+
+{{/ layout}}
