+++ views/prototype/photo/error.html
@@ -1,0 +1,21 @@
+{{< layout}}
+{{$propositionHeader}}{{/propositionHeader}}
+
+{{$pageTitle}}Code error{{/pageTitle}}
+
+{{$header}}
+    <h1>We can't find your code</h1>
+{{/header}}
+
+{{$content}}
+<p>Check your code and try again. You need to enter the code exactly as it’s written.
+
+<p>You entered: ab.de/1fG2h3i</p>
+
+<p>If this keeps happening, contact the company that provided your digital photo.</p>
+
+<a href="/prototype/intro/get-photo-code" class="button">Try again</a><br/><br/>
+
+{{/content}}
+
+{{/ layout}}
