+++ views/prototype/photo/get-photo-code.html
@@ -1,0 +1,47 @@
+{{< layout}}
+
+{{$pageTitle}}Upload your photo{{/pageTitle}}
+
+{{$header}}
+    <h1>Retrieve your digital photo</h1>
+{{/header}}
+
+{{$content}}
+
+    {{< hmpo-partials-form}}
+    {{$form}}
+        <div id="photo-code-group" class="form-group">
+            <h2 class="heading-medium">Photo code</h2>
+            <p class="form-hint" id="dob-hint" style="margin-bottom:10px">This is case-sensitive, for example abc.de/1fG2h3i4</p>
+
+            <input type="text" name="photo-code-sld" id="photo-code-sld" class="form-control input-code photo-code-long" aria-required="true">
+            <span id="reference-prefix" class="prefix" style="font-weight:bold;">.</span>
+            <input type="text" name="photo-code-tld" id="photo-code-tld" class="form-control input-code photo-code" aria-required="true">
+            <span id="reference-prefix" class="prefix" style="font-weight:bold;">/</span>
+            <input type="text" name="photo-code-path" id="photo-code-path" class="form-control input-code photo-code-long" aria-required="true">
+        </div>
+        <br>
+
+        {{#input-submit}}{{/input-submit}}
+
+    {{/form}}
+    {{/ hmpo-partials-form}}
+
+{{/content}}
+
+{{$sidebar}}
+<div class="column-one-third">
+    <br>
+    <br>
+    <div class="sidebar related">
+        <div class="photo-preview-inner">
+            <h2 class="heading-medium">Where to find your code</h2>
+            <div class="photo">
+                <img src="../../public/images/printed-photo-code.png" width="250" height="285" alt="">
+            </div>
+        </div>
+    </div>
+</div>
+{{/sidebar}}
+
+{{/ layout}}
