+++ views/prototype/photo/photo-retrieved.html
@@ -1,0 +1,49 @@
+{{< layout}}
+
+{{$pageTitle}}Photo retrieved{{/pageTitle}}
+
+{{$header}}
+    <h1>Your digital passport photo</h1>
+{{/header}}
+
+{{$content}}
+    <!--
+        <h2>Photo passed first check</h2>
+        <p>Your photo has been automatically checked for common mistakes.</p>
+        <p>We’ll still check it manually and we could still ask for another if something doesn't look right.</p>
+
+        <div class="panel panel-border-wide">
+            <p>
+              If you're not happy with your photo, <a href="/uploadphoto">upload a new one now</a>.
+            </p>
+        </div>
+-->
+
+        <p>Your photo must be taken in the last month.</p> 
+
+        <a href="/prototype/uploadphoto" class="button">Use this photo</a><br/><br/>
+        <a href="you-need-a-photo">I don't want to use this photo</a></p>
+
+         <p><div role="note" aria-label="Information" class="panel panel-border-wide application-notice info-notice">
+          We keep all photos for up to 30 days in line with our <a href="">privacy policy</a>.
+        </p>
+      </div>
+
+
+{{/content}}
+
+{{$sidebar}}
+<div class="column-one-third">
+    <div class="sidebar related">
+      <div class="photo-preview-inner">
+          <h2 class="heading-medium">Your photo</h2>
+          <div class="photo">
+              <img src="../../public/images/thumbnail.jpeg" alt="Your uploaded photo" title="Your uploaded photo" width="150">
+          </div>
+      </div>
+
+    </div>
+</div>
+{{/sidebar}}
+
+{{/ layout}}
