+++ views/prototype/photo/you-need-a-photo-overseas.html
@@ -1,0 +1,40 @@
+{{< layout}}
+
+{{$main-content}}
+
+<div class="column-full">
+    <div class="flash-card">
+        <header>
+          <h1>You need a digital passport photo</h1>
+        </header>
+        <div class="grid-row">
+            <div class="column-half">
+              <p>You can get a friend to take a digital photo then upload it - we'll show you how.</p>
+              <br>
+              <a href="/change_of_name_180122/intro/choose-photo-method-overseas" class="button">Continue</a>
+
+            </div>
+            <div class="column-half">
+              <img src="../../public/images/Photo_methods_@2x.png" alt="">
+            </div>
+        </div>
+        <div class="grid-row">
+            <div class="column-two-thirds">
+              <details>
+                    <summary><span class="summary">I have printed passport photos</span></summary>
+                    <div class="panel panel-border-narrow">
+                        You need to post 2 printed photos with your old passport. They must be taken within the last month.</p>
+                      <p>We’ll check your photos. Your application will be delayed if we have to ask for different ones.</p>
+                      <p><a href="/../change_of_name_180122/renew-without-photo">Apply with printed photos</a></p>
+                    </div>
+                </details>
+            </div>
+            <div class="column-one-third">
+            </div>
+        </div>
+    </div>
+</div>
+
+{{/main-content}}
+
+{{/ layout}}
