+++ views/prototype/photo/you-need-a-photo.html
@@ -1,0 +1,45 @@
+{{< layout}}
+{{$main-content}}
+{{< hmpo-partials-form}}
+  {{$form}}
+<div class="column-full">
+  <div class="flash-card">
+
+    <div class="grid-row">
+      <div class="column-two-thirds">
+        <header>
+          <h1 style="margin-bottom:30px">Get your digital passport photo</h1>
+        </header>
+
+        <p>You can choose from 3 ways to get your digital photo.</p>
+
+        <h2>Get a friend to take one</h2>
+        <p>They can use a phone, tablet or digital camera. We'll show you how to take a good passport photo and upload it.</p>
+
+        <h2>Go to a photo shop</h2>
+        <p>At a shop, you'll get a digital copy of your photo. Some shops can also give you a photo code. We'll show you how
+          to use the code to retrieve your digital photo.</p>
+
+        <h2>Go to a booth</h2>
+        <p>From some booths in the UK and some booths abroad, you'll get a photo code with your printed photos. We'll show you
+          how to use the code to retrieve your digital photo.</p>
+
+        {{#input-submit}}{{/input-submit}}
+
+        <div class="panel panel-border-wide">
+          <p>The photo code is a new service and may not be available yet from photo booths or shops in your area.</p>
+        </div>
+
+      </div>
+
+      <div class="column-one-third">
+        <img src="../../public/images/Photo_methods_2@2x.png" alt="" style="height:300px">
+      </div>
+    </div>
+
+  </div>
+</div>
+{{/form}}
+{{/ hmpo-partials-form}}
+{{/main-content}}
+{{/ layout}}
